+++ CHANGELOG.md
@@ -11,6 +11,8 @@ CHANGELOG for FlatCAM beta
 
 - added new functionality: centering in origin for a selection of objects
 - for all centering functionality, now if the moved object is of 'geometry' kind it will have the source_file attribute updated with DXF code
+- modified the centering in origin functionality to include a selection of choices in a dialog
+- fixed Locate in Object functionality to work in the situations where there are negative bounds values
 
 13.12.2020